+++ java-operators/test/com/baeldung/java/operators/AssignmentTest.java
@@ -1,0 +1,18 @@
+package com.baeldung.java.operators;
+
+import static org.junit.Assert.assertEquals;
+
+import org.junit.jupiter.api.Test;
+
+public class AssignmentTest{
+
+	
+	@Test
+	public void whenUseTheOperatorEqualTo() {
+		int grade = 3;
+		
+		assertEquals(grade, 3);
+	
+	}
+	
+}
