+++ java-operators/test/com/baeldung/java/operators/ConditionalTest.java
@@ -1,0 +1,69 @@
+package com.baeldung.java.operators;
+
+import static org.junit.Assert.assertEquals;
+
+import org.junit.jupiter.api.Test;
+
+public class ConditionalTest {
+
+	
+	@Test
+	public void whenUseTheOperatorConditionalAnd() {
+		
+		int years = 25;
+		boolean driverLicence = true;
+		String result = "";
+		
+		if((years >= 21 && driverLicence)) {
+			result = "Successful Candidate.";
+		}else {
+			result = "Failed Candidate.";
+		}
+		
+		assertEquals(result, "Successful Candidate.");
+		
+	}
+	
+	@Test
+	public void whenUseTheOperatorConditionalAndTernary() {
+		
+		int years = 25;
+		boolean driverLicence = true;
+		String result = (years >= 21 && driverLicence) ?  "Successful Candidate." : "Successful Candidate.";
+		
+		assertEquals(result, "Successful Candidate.");
+		
+	}
+	
+	@Test
+	public void whenUseTheOperatorConditionalOr() {
+		
+		int years = 25;
+		boolean driverLicence = false;
+		String result = "";
+		
+		if((years >= 21 || driverLicence)) {
+			result = "Successful Candidate.";
+		}else {
+			result = "Failed Candidate.";
+		}
+		
+		assertEquals(result, "Successful Candidate.");
+		
+	}
+	
+	@Test
+	public void whenUseTheOperatorConditionalOrTernary() {
+		
+		int years = 25;
+		boolean driverLicence = false;
+		String result = (years >= 21 || driverLicence) ?  "Successful Candidate." : "Successful Candidate.";
+		
+		assertEquals(result, "Successful Candidate.");
+		
+	}
+	
+
+	
+		
+}
