--- java-operators/test/com/baeldung/operators/AssignmentTest.java
@@ -1,16 +1,0 @@
-package com.baeldung.operators;
-
-import static org.junit.Assert.assertEquals;
-
-import org.junit.jupiter.api.Test;
-
-public class AssignmentTest{
-
-	
-	@Test
-	public void greaterThanorEqualToTest() {
-		int grade = 3;
-		assertEquals(grade, 3);
-	}
-	
-}
